requirements.txt
openai>=1.0.0
pandas
numpy
scikit-learn
pytest
python-dotenv